+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "smcloudstore-builder",
+  "name": "@smcloudstore/builder",
   "private": true,
-  "version": "0.1.0",
+  "version": "0.0.1",
   "scripts": {
     "bootstrap": "npx lerna bootstrap --hoist",
     "compile": "npm run tsc",
@@ -17,17 +17,16 @@
   "author": "Alessandro Segala",
   "license": "MIT",
   "homepage": "https://github.com/ItalyPaleAle/SMCloudStore",
-  "dependencies": {
-  },
+  "dependencies": {},
   "devDependencies": {
     "@types/google-cloud__storage": "^1.7.0",
     "@types/minio": "^6.0.2",
-    "@types/node": "^10.5.7",
+    "@types/node": "^10.5.8",
     "coveralls": "^3.0.2",
     "digest-stream": "^2.0.0",
     "eslint": "^5.3.0",
     "istanbul": "v1.1.0-alpha.1",
-    "lerna": "^3.0.0-rc.0",
+    "lerna": "^3.0.2",
     "mocha": "^5.2.0",
     "mocha-lcov-reporter": "^1.3.0",
     "randomstring": "^1.1.5",
