+++ packages/aws-s3/package.json
@@ -1,11 +1,11 @@
 {
-  "name": "smcloudstore-minio",
-  "version": "0.1.0",
+  "name": "@smcloudstore/aws-s3",
+  "version": "0.0.1",
   "description": "> TODO: description",
   "author": "Alessandro Segala <ItalyPaleAle@users.noreply.github.com>",
-  "homepage": "https://github.com/ItalyPaleAle/SMCloudStore/tree/master/packages/smcloudstore-minio",
+  "homepage": "https://github.com/ItalyPaleAle/SMCloudStore/tree/master/packages/aws-s3",
   "license": "MIT",
-  "main": "dist/MinioProvider.js",
+  "main": "dist/AwsS3Provider.js",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/ItalyPaleAle/SMCloudStore.git"
@@ -17,6 +17,6 @@
     "url": "https://github.com/ItalyPaleAle/SMCloudStore/issues"
   },
   "dependencies": {
-    "smcloudstore-generic-s3": "^0.1.0"
+    "@smcloudstore/generic-s3": "^0.0.1"
   }
 }
