+++ packages/aws-s3/src/AwsS3Provider.ts
@@ -1,7 +1,7 @@
 'use strict'
 
+import GenericS3Provider from '@smcloudstore/generic-s3'
 import Minio from 'minio'
-import GenericS3Provider from 'smcloudstore-generic-s3'
 
 /**
  * Connection options for an AWS S3 provider.
