+++ packages/aws-s3/tsconfig.json
