+++ packages/azure-storage/README.md
@@ -1,0 +1,1 @@
+# @smcloudstore/azure-storage
