+++ packages/core/README.md
@@ -1,0 +1,1 @@
+# @smcloudstore/core
