+++ packages/generic-s3/README.md
@@ -1,0 +1,1 @@
+# @smcloudstore/generic-s3
