+++ packages/google-cloud-storage/README.md
@@ -1,0 +1,1 @@
+# @smcloudstore/google-cloud-storage
