--- packages/smcloudstore-aws-s3/README.md
@@ -1,1 +1,0 @@
-# smcloudstore-aws-s3
