--- packages/smcloudstore-azure-storage/README.md
@@ -1,1 +1,0 @@
-# smcloudstore-azure-storage
