--- packages/smcloudstore-core/README.md
@@ -1,1 +1,0 @@
-# smcloudstore-core
