--- packages/smcloudstore-generic-s3/README.md
@@ -1,1 +1,0 @@
-# smcloudstore-generic-s3
