--- packages/smcloudstore-minio/README.md
@@ -1,1 +1,0 @@
-# smcloudstore-minio
