+++ scripts/bootstrap.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+npm install
+npx lerna bootstrap --hoist
