+++ test/GoogleCloudStorateProvider.test.js
@@ -15,7 +15,7 @@ TestSuite('google-cloud-storage', testSuiteOptions)
 
 // Add custom, provider-specific tests
 describe('Provider-specific tests for google-cloud-storage', function() {
-    const Provider = require('../packages/smcloudstore-google-cloud-storage')
+    const Provider = require('../packages/google-cloud-storage')
 
     it('constructor', function() {
         assert.throws(() => {
